==> src/Part1_WebScraping.ipynb <==
# See getting-started-selenium to install the requirements (*)
import csv
import time
import pandas
from selenium import webdriver

# Select location and neighborhood from get restaurants, and copy the start page:
neighborhood = "Arguelles"
start_page = "https://www.tripadvisor.es/Restaurants-g187514-Madrid.html"

# Select maximun number of links to be scraped (max num of restaurant per page) and maximun pages to go inside 
# ([1][2]...)
MAX_LINK_NUM = 30
MAX_PAGE_NUM = 1

# Open up a Firefox browser and navigate to wep page.
driver = webdriver.Firefox()
driver.get(start_page)
select = driver.find_elements_by_xpath('//div[@class="filterAddOn"]/div[text()="Más"]')
select[1].click()
select2 = driver.find_elements_by_xpath('//div[@class="label"]')
select2[1].click()
driver.find_element_by_xpath('//div[@class="applyButton ui_button primary"]').click()
time.sleep(5)

list_rest =[]
i=0
l=0
if i < MAX_PAGE_NUM:
    i=i+1
    links = driver.find_elements_by_xpath('//a[@class="property_title"]')
    time.sleep(5)
    num_links = len(links)
    
    for i in range(num_links):
        if (l<MAX_LINK_NUM):
            l=l+1
            links[i].click()
            time.sleep(5)
        
            # Webdriver searches for elements in the new page
            driver.switch_to_window(driver.window_handles[1])
        
            # Extract lists of characteristics based on xpath
            name = driver.find_element_by_xpath('//h1[@class="heading_title"]')
            towrite_n=name.text.encode('utf8', 'replace')
            puntuation = driver.find_element_by_xpath('//div[@class="rating"]/span[@class="overallRating"]')
            towrite_p=puntuation.text.encode('utf8', 'replace')
            address = driver.find_element_by_xpath('//span[@class="street-address"]')
            towrite_a=address.text.encode('utf8', 'replace')
            locality = driver.find_element_by_xpath('//span[@class="locality"]')
            towrite_l=locality.text.encode('utf8', 'replace')
            description = driver.find_element_by_xpath('//div[@class="text"]')
            towrite_d=description.text.encode('utf8', 'replace')
            
            mostrar = driver.find_element_by_xpath('//p[@class="partial_entry"]/span[text()="Más"]').click()
           
            time.sleep(5)
            
            comments_title = driver.find_elements_by_xpath('//span[@class="noQuotes"]')
            comments_text= driver.find_elements_by_xpath('//div[@class="wrap"]/div[@class="prw_rup prw_reviews_text_summary_hsx"]')
            comments_puntuation= driver.find_elements_by_xpath('//div[@class="rating reviewItemInline"]/span[contains(@class,"ui_bubble_rating")]')
            
            # Print out all of the restaurant, comment, puntuation...on current page
            num_page_url = len(comments_title)
            
            for i in range (num_page_url):
                towrite_t=comments_title[i].text.encode('utf8', 'replace').replace('\n', ' ')
                towrite_c=comments_text[i].text.encode('utf8', 'replace').replace('\n', ' ')
                towrite_cp=comments_puntuation[i].get_attribute("class")
                dataframe ={}
                dataframe["rest_name"] = towrite_n
                dataframe["rest_puntuation"] = towrite_p
                dataframe["rest_description"] = towrite_d
                dataframe["locality"] = towrite_l
                dataframe["comment_title"] = towrite_t
                dataframe["comment"] = towrite_c
                dataframe["comment_puntuation"] = towrite_cp
                
                list_rest.append(dataframe)
            list_rest
        
            # Close current page
            driver.close()
            driver.switch_to_window(driver.window_handles[0])
            time.sleep(3) 
            
df = pandas.DataFrame(list_rest)
df.to_csv("Arguelles.csv",index=False)  

# Clean up (close browser once task is completed).
driver.quit()                            

(*)getting-started-selenium:
- On Ubuntu/Debian systems, this will install pip (and dependencies) and then install the Selenium Python bindings from PyPI:
  - sudo apt-get install python-pip
  - sudo pip install selenium
- Selenium requires a driver to interface with the chosen browser. Firefox, for example, requires geckodriver.
